Default to shallow clone/fall back to regular, don't plug vim-plug:
vim-plug repo is pretty big (8.8M at the time of this commit),
so it makes sense to only fetch the HEAD of master (or cURL
using raw.githubusercontent.com if available). If however shallow
clone is not supported due to proxying/mirroring, regular git clone
will be used.

Maintaining vim-plug as a plugin can be a bad idea for several reasons,
please see the vim-plug FAQ:

https://github.com/junegunn/vim-plug/wiki/faq#shouldnt-vim-plug-update-itself-on-plugupdate-like-vundle